Converted example 14 to Vite. Need to update use-local-storage as its broken now.
