+++ example-14-localstorage-02/README.md
@@ -2,9 +2,9 @@
 This project contains two examples showing how we can integrate the use of local storage (`localStorage`) into our React apps.
 
 ## useLocalStorage() hook
-The first example shows a custom hook we have written - (useLocalStorage)[./src/useLocalStorage.js]. The hook essentially wraps a call to `useState()` - but adds the side-effect (using `useEffect()`) of saving the stateful value to `localStorage` under a given `key` whenever it is updated.
+The first example shows a custom hook we have written - [useLocalStorage](./src/useLocalStorage.js). The hook essentially wraps a call to `useState()` - but adds the side-effect (using `useEffect()`) of saving the stateful value to `localStorage` under a given `key` whenever it is updated.
 
-Its usage can be seen in `Counter.js`(./src/Counter.js). The benefit of this approach is that it is relatively simple and we can see what's going on - we have achieved `localStorage` integration with only a few lines of code and have encapsulated it in a custom hook to promote code reuse. The drawback is that the behaviour isn't as expected if we try to share the value with another component. When one component updates the value in local storage, that change will not automatically cause a re-render of any component other than the one which changed the value.
+Its usage can be seen in [`Counter.jsx`](./src/Counter.jsx). The benefit of this approach is that it is relatively simple and we can see what's going on - we have achieved `localStorage` integration with only a few lines of code and have encapsulated it in a custom hook to promote code reuse. The drawback is that the behaviour isn't as expected if we try to share the value with another component. When one component updates the value in local storage, that change will not automatically cause a re-render of any component other than the one which changed the value.
 
 To get around this issue, we could combine our `useLocalStorage()` function with React's Context mechanism as seen in a previous example. Something like this:
 
@@ -36,6 +36,8 @@ function CounterViewer() {
 }
 ```
 
+<!-- NOTE: The below is NOT working with React 18, as the library is no longer compatible. There may be another option available. If I find one I will update this README & example. -->
+<!--
 Alternatively, we could use a ready-made third-party solution: the `npm` package `use-persisted-state`.
 
 ## use-persisted-state package
@@ -53,3 +55,4 @@ We can see its usage in [LoginPage.js](./src/LoginPage.js) and [UserInfoPage.js]
 As we can see, we were able to use our knowledge of `useState()`, `useEffect()`, and `localStorage` to quickly put together a custom hook which is sufficient for many of our needs. And we were able to come up with an addition using React's Context API if we need to share and modify the same local storage value on multiple components.
 
 However, as we have also seen, a third-party package already exists- `use-persisted-state` - which offers the same functionality plus more. This is often true when developing modern software, but *especially* in the web development domain - it is often the case that someone will have solved a particular problem before. It is perfectly OK (given appropriate licensing & credit where credit is due) to integrate third-party solutions into your own projects to help you save time and energy, allowing you to devote your effort to the specific requirements of your project.
+-->
