+++ example-14-localstorage-02/src/LoginPage.jsx
@@ -1,0 +1,19 @@
+// import createPersistedState from 'use-persisted-state';
+
+// const useAuthState = createPersistedState('auth');
+
+// /**
+//  * Renders a "login page" containing login / logout buttons. Clicking either of these buttons causes the setUser() function,
+//  * which has been obtained from local storage, to be called.
+//  */
+// export default function LoginPage() {
+
+//     const [user, setUser] = useAuthState();
+
+//     return (
+//         <div>
+//             <button onClick={() => setUser({ username: 'Bob' })}>Log in</button>
+//             <button onClick={() => setUser(null)}>Log out</button>
+//         </div>
+//     );
+// }
